% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-window.R
\name{apply_window}
\alias{apply_window}
\title{Filter out transactions older than window}
\usage{
apply_window(input, window)
}
\arguments{
\item{input}{data frame with transactions for export}

\item{window}{integer variable containing maximum age of transaction in days}
}
\value{
output data frame with filtered transactions
}
\description{
Remove any transactions older than window days from the input data frame
}
\examples{
apply_window(input, 90)
}
