% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-amex.R
\name{convert_amex}
\alias{convert_amex}
\title{Convert American Express qfx file for GnuCash}
\usage{
convert_amex(file, outfile = NULL, overwrite = TRUE, window)
}
\arguments{
\item{file}{Character variable containing input file in qfx format}

\item{outfile}{Character variable containing output file name or NULL to
create a sensible name in the same directory as the input file}

\item{overwrite}{logical variable to control whether an existing output file
of the same name will be overwritten}

\item{window}{arugument for compatibility only - is ignore}
}
\description{
This is actually a very trivial conversion as the qfx file is nicely formatted. However,
the currency is currently specified as EUR for Swedish accounts so this function just changes
that into SEK, which allows the file to be loaded into GnuCash
}
\examples{
convert_amex("~/Desktop/activity.qfx")
convert_amex("activity.qfx", outfile = "NEW.qfx", overwrite = FALSE)

}
