% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-circlek.R
\name{convert_circlek}
\alias{convert_circlek}
\title{Convert Circle K Excel transaction file info CSV for GnuCash}
\usage{
convert_circlek(file, outfile = NULL, overwrite = TRUE)
}
\arguments{
\item{file}{Character variable containing input file name (XLSXfile)}

\item{outfile}{Character variable containing output file name or NULL to
create a sensible name in the same directory as the input file}

\item{overwrite}{logical variable to control whether an existing output file
of the same name will be overwritten}
}
\description{
This extracts relevant columns and writes CSV file
}
\examples{
convert_circlek("~/Desktop/transactions.xlsx")
}
