% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-wise.R
\name{convert_wise}
\alias{convert_wise}
\title{Convert a Wise statement to a CSV file suitable for import into GnuCash}
\usage{
convert_wise(file, outfile = NULL, overwrite = TRUE)
}
\arguments{
\item{file}{Character variable containing input file name (CSV file)}

\item{outfile}{Character variable containing output file name or NULL to
create a sensible name in the same directory as the input file}

\item{overwrite}{logical variable to control whether an existing output file
of the same name will be overwritten}
}
\description{
You need to export a statement from the WISE website (and not download transactions).
This is found under your name on the top right. Use Create a statement, choose the dates
and currency and select CSV format and English (UK) as the Statement language. Do not select
Display transactions with fees separately. Then click Generate and Download the resulting
statement.
}
\examples{
convert_wise("~/Desktop/statement_15924_SEK_2026-01-01_2026-05-29.csv", overwrite = FALSE)
convert_wise("~/Desktop/statement_15924_SEK_2026-01-01_2026-05-29.csv", outfile = "TMP.csv")
}
