% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-statements.R
\name{process_statements}
\alias{process_statements}
\title{Convert bank statements in specified directory}
\usage{
process_statements(path = ".", window = 90)
}
\arguments{
\item{path}{character string containg directory to process}

\item{window}{integer containing maximum age of transactions to import}
}
\description{
Convert bank statements in specified directory
}
\examples{
process_statements()
}
