% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-statements.R
\name{process_statements}
\alias{process_statements}
\title{Convert bank statements in specified directory}
\usage{
process_statements(path = ".")
}
\arguments{
\item{path}{character str4ing containg directory to process}
}
\description{
Convert bank statements in specified directory
}
\examples{
process_statements()
}
